+++ Source/Queries/IOperation.php
@@ -12,7 +12,6 @@ interface IOperation
 {
     const APPLY = 0;
     const JOIN_APPLY = 1;
-    const EQUALITY_JOIN_APPLY = 2;
     const REMOVE_VALUES = 3;
     const ADD_VALUES = 4;
     const CLEAR = 5;